Moved test-utility files to their dedicated folder. Updated useForm to remove unnecessary env checks.
